Added a TimesheetViewer controller and accompanying view to handle the rendering of the timesheet report.
As a side note, a strange anomaly was encountered related to the in memory database model. The TimesheetEntry referenced by the Timesheet held by the DbSet was being lost once stored which prevented the report from being rendered. A temporary workaround has been employed, but will be revoked once the underlying issue is further understood.