Attached an action to the pre-existing Export Timesheets button to fire a method of the Timesheet controller.
This method exports the data to a local CSV file (using an imported nuget package) then redirects to the main Index view.

The reality is that a case where pre-existing functionality is already available (in this case a 3rd party nuget package), writing a dedicated CSV helper class is simply unnecessary and the time can be better spent elsewhere.